# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "电信客服 Agent v1",
    "version": "0.1.0"
  },
  "paths": {
...
    "/agents/conversations/{conversation_id}/tickets": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "List Conversation Tickets",
        "description": "坐席查询会话所属工单列表（US-23，B12 issue #44 AC3）。\n\nactive-chat 右栏「当前工单」数据源；会话可见性同消息历史（仅 handed_off）。",
        "operationId": "list_conversation_tickets_agents_conversations__conversation_id__tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketOut"
                  },
                  "title": "Response List Conversation Tickets Agents Conversations  Conversation Id  Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/tickets": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "Create Agent Ticket",
        "description": "坐席为转接会话创建工单（US-23，B12 issue #44 AC3）。\n\ncreator_type=agent、creator_id=坐席；customer_id 关联会话所属客户（访客会话\n为 None 仅落联系方式）。仅 handed_off 转接中的会话可建单（坐席在 active-chat\n右栏「创建工单」Modal 提交）。",
        "operationId": "create_agent_ticket_agents_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/tickets": {
      "get": {
        "tags": [
          "ticket"
        ],
        "summary": "List Tickets",
        "description": "当前客户的工单列表（未认证 401 由 CurrentCustomer 守卫）。",
        "operationId": "list_tickets_tickets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TicketOut"
                  },
                  "type": "array",
                  "title": "Response List Tickets Tickets Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "ticket"
        ],
        "summary": "Create",
        "description": "创建工单（办理类/工单类），customer 关联当前认证客户，默认 pending 入队。",
        "operationId": "create_tickets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "ticket"
        ],
        "summary": "Get Ticket",
        "description": "工单详情；不存在或不属于当前客户 → 404（不泄露存在性）。",
        "operationId": "get_ticket_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ticket"
        ],
        "summary": "Update Status",
        "description": "工单状态流转：按 ticket_type 双状态机校验，非法 → 422 状态不变。\n\n成功后推送 WS ticket.update；触发通知态（办理生效/失败、工单派单/关闭）\n推送 notification.push 站内通知。",
        "operationId": "update_status_tickets__ticket_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notifications": {
      "get": {
        "tags": [
          "notification"
        ],
        "summary": "List Notifications",
        "description": "当前客户的站内通知列表（US-14，UI-C-4 通知预览条冷启动数据源）。\n\nB13（issue #53 AC2）：notifications 表 + WS notification.push（B7）已有，\n本端点补 REST 读取——页面打开前产生的通知可获取；按 created_at 倒序\n（同 created_at 按 id 倒序），含 read 未读标记（前端预览条过滤未读）。\n仅返回当前客户的通知（主体隔离）；未认证 401 由 CurrentCustomer 守卫。",
        "operationId": "list_notifications_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationOut"
                  },
                  "type": "array",
                  "title": "Response List Notifications Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotificationOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ticket_id": {
            "type": "integer",
            "title": "Ticket Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "read": {
            "type": "boolean",
            "title": "Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ticket_id",
          "message",
          "read",
          "created_at"
        ],
        "title": "NotificationOut",
        "description": "站内通知（GET /notifications 响应，US-14 通知预览条冷启动数据源）。\n\n镜像 Notification 模型（ticket_id/message/read/created_at）；\nread 未读标记供前端预览条过滤未读。"
      },
...
      "TicketCreate": {
        "properties": {
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id",
            "description": "所属会话"
          },
          "ticket_type": {
            "type": "string",
            "enum": [
              "transaction",
              "ticketing"
            ],
            "title": "Ticket Type",
            "description": "工单类型：transaction/ticketing"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "工单内容（办理事项描述 / 报修描述）"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "ticket_type",
          "content"
        ],
        "title": "TicketCreate",
        "description": "创建工单请求（REST 客户侧；customer 自动取当前认证客户）。"
      },
      "TicketOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "conversation_id": {
            "type": "integer",
            "title": "Conversation Id"
          },
          "ticket_type": {
            "type": "string",
            "title": "Ticket Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "creator_type": {
            "type": "string",
            "title": "Creator Type"
          },
          "creator_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "conversation_id",
          "ticket_type",
          "status",
          "content",
          "customer_id",
          "contact_name",
          "contact_phone",
          "creator_type",
          "creator_id",
          "created_at"
        ],
        "title": "TicketOut",
        "description": "工单详情/列表项。customer 关联允许 null（Visitor 创建时仅联系方式）。"
      },
      "TicketStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "目标状态（两状态机合法状态名）"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TicketStatusUpdate",
        "description": "状态流转请求（PATCH /tickets/{id}，按工单类型状态机校验）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}